+++ changelog.md
@@ -4,6 +4,8 @@
 * LLT-5004: Wireguard go oscillation
 * LLT-1113: Add no-link detection mechanism
 * LLT-4983: Expose telio_generate_stack_panic method
+* LLT-4509: Enable windows arm64 build.
+* LLT-5007: Bump moose tracker to v3.0.0-libtelioApp
 
 <br>